+++ examples/nextjs/app/edge/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { env } from "~/env.mjs";
+import { env } from "~/env";
 
 export const runtime = "edge";