+++ examples/nextjs/next.config.js
@@ -1,0 +1,6 @@
+import "./app/env.js";
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {};
+
+export default nextConfig;